+++ src/interfaces/index.ts
@@ -7,6 +7,8 @@ export * from './flow-job';
 export * from './job-json';
 export * from './jobs-options';
 export * from './keep-jobs';
+export * from './metrics-options';
+export * from './metrics';
 export * from './parent-command';
 export * from './parent-message';
 export * from './parent';