+++ tests/test_connection.ts
@@ -89,6 +89,20 @@ See https://https://github.com/OptimalBits/bull/issues/1873`);
     });
   });
 
+  describe('when host belongs to Upstash', async () => {
+    it('throws an error', async () => {
+      const opts = {
+        connection: {
+          host: 'https://upstash.io',
+        },
+      };
+
+      expect(() => new QueueBase(queueName, opts)).to.throw(
+        'BullMQ: Upstash is not compatible with BullMQ.',
+      );
+    });
+  });
+
   it('should recover from a connection loss', async () => {
     let processor;