+++ tests/test_worker.ts
@@ -7,9 +7,10 @@ import { v4 } from 'uuid';
 import {
   Queue,
   QueueEvents,
-  Job,
-  Worker,
   QueueScheduler,
+  Job,
+  UnrecoverableError,
+  Worker,
 } from '../src/classes';
 import { KeepJobs, JobsOptions } from '../src/interfaces';
 
@@ -1774,8 +1775,8 @@ describe('workers', function () {
       await queueScheduler.close();
     });
 
-    describe('when providing a way to execute step jobs', () => {
-      it('should retry a job after a delay if a fixed backoff is given, keeping the current step', async function () {
+    describe('when UnrecoverableError is throw', () => {
+      it('moves job to failed', async function () {
         this.timeout(8000);
 
         const queueScheduler = new QueueScheduler(queueName, { connection });
@@ -1784,28 +1785,82 @@ describe('workers', function () {
         const worker = new Worker(
           queueName,
           async job => {
-            const initialStep = 'initialStep';
-            const secondStep = 'secondStep';
-            const finishStep = 'finishStep';
+            if (job.attemptsMade < 2) {
+              throw new Error('Not yet!');
+            }
+            if (job.attemptsMade < 3) {
+              throw new UnrecoverableError('Unrecoverable');
+            }
+          },
+          { connection },
+        );
+
+        await worker.waitUntilReady();
+
+        const start = Date.now();
+        await queue.add(
+          'test',
+          { foo: 'bar' },
+          {
+            attempts: 3,
+            backoff: 1000,
+          },
+        );
+
+        await new Promise<void>(resolve => {
+          worker.on(
+            'failed',
+            after(2, (job: Job, error) => {
+              const elapse = Date.now() - start;
+              expect(error.name).to.be.eql('UnrecoverableError');
+              expect(error.message).to.be.eql('Unrecoverable');
+              expect(elapse).to.be.greaterThan(1000);
+              expect(job.attemptsMade).to.be.eql(2);
+              resolve();
+            }),
+          );
+        });
+
+        await worker.close();
+        await queueScheduler.close();
+      });
+    });
+
+    describe('when providing a way to execute step jobs', () => {
+      it('should retry a job after a delay if a fixed backoff is given, keeping the current step', async function () {
+        this.timeout(8000);
+
+        const queueScheduler = new QueueScheduler(queueName, { connection });
+        await queueScheduler.waitUntilReady();
+
+        enum Step {
+          Initial,
+          Second,
+          Finish,
+        }
+
+        const worker = new Worker(
+          queueName,
+          async job => {
             let step = job.data.step;
-            while (step !== finishStep) {
+            while (step !== Step.Finish) {
               switch (step) {
-                case initialStep: {
+                case Step.Initial: {
                   await job.update({
-                    step: secondStep,
+                    step: Step.Second,
                   });
-                  step = secondStep;
+                  step = Step.Second;
                   break;
                 }
-                case secondStep: {
+                case Step.Second: {
                   if (job.attemptsMade < 3) {
                     throw new Error('Not yet!');
                   }
                   await job.update({
-                    step: finishStep,
+                    step: Step.Finish,
                   });
-                  step = finishStep;
-                  return 'finished';
+                  step = Step.Finish;
+                  return Step.Finish;
                 }
                 default: {
                   throw new Error('invalid step');
@@ -1823,7 +1878,7 @@ describe('workers', function () {
         const start = Date.now();
         await queue.add(
           'test',
-          { step: 'initialStep' },
+          { step: Step.Initial },
           {
             attempts: 3,
             backoff: 1000,
@@ -1834,7 +1889,7 @@ describe('workers', function () {
           worker.on('completed', job => {
             const elapse = Date.now() - start;
             expect(elapse).to.be.greaterThan(2000);
-            expect(job.returnvalue).to.be.eql('finished');
+            expect(job.returnvalue).to.be.eql(Step.Finish);
             expect(job.attemptsMade).to.be.eql(3);
             resolve();
           });
@@ -1842,6 +1897,121 @@ describe('workers', function () {
 
         await worker.close();
         await queueScheduler.close();
+      });
+
+      describe('when creating children at runtime', () => {
+        it('should wait children as one step of the parent job', async function () {
+          this.timeout(8000);
+          const parentQueueName = `parent-queue-${v4()}`;
+          const parentQueue = new Queue(parentQueueName, { connection });
+
+          const queueScheduler = new QueueScheduler(parentQueueName, {
+            connection,
+          });
+          await queueScheduler.waitUntilReady();
+
+          enum Step {
+            Initial,
+            Second,
+            Third,
+            Finish,
+          }
+
+          const worker = new Worker(
+            parentQueueName,
+            async (job, token) => {
+              let step = job.data.step;
+              while (step !== Step.Finish) {
+                switch (step) {
+                  case Step.Initial: {
+                    await queue.add(
+                      'child-1',
+                      { foo: 'bar' },
+                      {
+                        parent: {
+                          id: job.id,
+                          queue: `${job.prefix}:${job.queueName}`,
+                        },
+                      },
+                    );
+                    await job.update({
+                      step: Step.Second,
+                    });
+                    step = Step.Second;
+                    break;
+                  }
+                  case Step.Second: {
+                    if (job.attemptsMade < 3) {
+                      throw new Error('Not yet!');
+                    }
+                    await queue.add(
+                      'child-2',
+                      { foo: 'bar' },
+                      {
+                        parent: {
+                          id: job.id,
+                          queue: `bull:${parentQueueName}`,
+                        },
+                      },
+                    );
+                    await job.update({
+                      step: Step.Third,
+                    });
+                    step = Step.Third;
+                    break;
+                  }
+                  case Step.Third: {
+                    const shouldWait = await job.moveToWaitingChildren(token);
+                    if (!shouldWait) {
+                      await job.update({
+                        step: Step.Finish,
+                      });
+                      step = Step.Finish;
+                      return Step.Finish;
+                    }
+                    break;
+                  }
+                  default: {
+                    throw new Error('invalid step');
+                  }
+                }
+              }
+            },
+            { connection },
+          );
+          const childrenWorker = new Worker(
+            queueName,
+            async () => {
+              await delay(100);
+            },
+            {
+              connection,
+            },
+          );
+          await childrenWorker.waitUntilReady();
+          await worker.waitUntilReady();
+
+          await parentQueue.add(
+            'test',
+            { step: Step.Initial },
+            {
+              attempts: 3,
+              backoff: 1000,
+            },
+          );
+
+          await new Promise<void>(resolve => {
+            worker.on('completed', job => {
+              expect(job.returnvalue).to.equal(Step.Finish);
+              resolve();
+            });
+          });
+
+          await worker.close();
+          await childrenWorker.close();
+          await parentQueue.close();
+          await queueScheduler.close();
+        });
       });
     });